[refactor-minimax] Fix imports
