+++ src/app/games/abalone/tests/AbaloneRules.spec.ts
@@ -4,7 +4,6 @@ import { FourStatePiece } from 'src/app/jscaip/FourStatePiece';
 import { HexaDirection } from 'src/app/jscaip/HexaDirection';
 import { Minimax } from 'src/app/jscaip/Minimax';
 import { Player } from 'src/app/jscaip/Player';
-import { GameStatus } from 'src/app/jscaip/Rules';
 import { RulesFailure } from 'src/app/jscaip/RulesFailure';
 import { RulesUtils } from 'src/app/jscaip/tests/RulesUtils.spec';
 import { AbaloneDummyMinimax } from '../AbaloneDummyMinimax';
@@ -12,6 +11,7 @@ import { AbaloneFailure } from '../AbaloneFailure';
 import { AbaloneState } from '../AbaloneState';
 import { AbaloneMove } from '../AbaloneMove';
 import { AbaloneLegalityInformation, AbaloneNode, AbaloneRules } from '../AbaloneRules';
+import { GameStatus } from 'src/app/jscaip/GameStatus';
 
 describe('AbaloneRules', () => {
