+++ src/app/games/quarto/QuartoRules.ts
@@ -1,4 +1,4 @@
-import { GameStatus, Rules } from '../../jscaip/Rules';
+import { Rules } from '../../jscaip/Rules';
 import { MGPNode } from 'src/app/jscaip/MGPNode';
 import { QuartoState } from './QuartoState';
 import { QuartoMove } from './QuartoMove';
@@ -14,6 +14,7 @@ import { QuartoFailure } from './QuartoFailure';
 import { MGPOptional } from 'src/app/utils/MGPOptional';
 import { MGPSet } from 'src/app/utils/MGPSet';
 import { CoordSet } from 'src/app/utils/OptimizedSet';
+import { GameStatus } from 'src/app/jscaip/GameStatus';
 
 /**
  * A criterion is a list of boolean sub-criteria, so three possible values: true, false, null.
