+++ src/app/games/tafl/TaflPieceAndControlMinimax.ts
@@ -2,7 +2,6 @@ import { Coord } from 'src/app/jscaip/Coord';
 import { Orthogonal } from 'src/app/jscaip/Direction';
 import { BoardValue } from 'src/app/jscaip/BoardValue';
 import { Player } from 'src/app/jscaip/Player';
-import { GameStatus } from 'src/app/jscaip/Rules';
 import { MGPMap } from 'src/app/utils/MGPMap';
 import { MGPSet } from 'src/app/utils/MGPSet';
 import { TaflPawn } from './TaflPawn';
@@ -12,6 +11,7 @@ import { SandwichThreat } from '../../jscaip/PieceThreat';
 import { TaflNode } from './TaflMinimax';
 import { CoordSet } from 'src/app/utils/OptimizedSet';
 import { assert } from 'src/app/utils/assert';
+import { GameStatus } from 'src/app/jscaip/GameStatus';
 
 export type TaflPieceAndControlMinimaxMetrics = {
     controlScore: number,
