+++ src/app/jscaip/tests/Rules.spec.ts
@@ -1,10 +1,11 @@
 /* eslint-disable max-lines-per-function */
 import { P4Move } from 'src/app/games/p4/P4Move';
 import { MGPNode } from '../MGPNode';
-import { GameStatus, Rules } from '../Rules';
+import { Rules } from '../Rules';
 import { GameStateWithTable } from '../GameStateWithTable';
 import { MGPOptional } from 'src/app/utils/MGPOptional';
 import { MGPValidation } from '../../utils/MGPValidation';
+import { GameStatus } from '../GameStatus';
 
 class MyAbstractState extends GameStateWithTable<number> {
